requirements.txt
yfinance>=0.2.28
pandas>=1.5.0
numpy>=1.23.0
plotly>=5.13.0
feedparser>=6.0.10
vaderSentiment>=3.3.2
requests>=2.28.0
xlsxwriter>=3.1.0
flask>=2.3.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
